refactor: utils로 함수 분리
